ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
anthropic==0.18.1
Pillow==10.2.0
python-dotenv==1.0.1
httpx==0.26.0
aiofiles==23.2.1
pydantic==2.5.3
